refactor(composer): fold the address parser into recipients
