--- desk/src/components/communication-area/addresses.ts
@@ -1,15 +1,0 @@
-import type { Recipient } from "@framework/ui/components/Composer/index.ts";
-
-// Bare addresses and the `Name <email>` form the timeline hands over.
-export function toRecipient(address: string): Recipient {
-  const match = address.match(/^\s*"?([^"<]*?)"?\s*<([^>]+)>\s*$/);
-  if (!match) return { email: address.trim() };
-  const label = match[1].trim();
-  return { email: match[2].trim(), ...(label && { label }) };
-}
-
-export function toRecipientList(
-  addresses: (string | undefined)[] = []
-): Recipient[] {
-  return addresses.filter((a): a is string => !!a).map(toRecipient);
-}
